chore: Cache node modules in CI workflow
This commit adds a step to the CI workflow to cache node modules. Caching the node modules will improve the build time by reusing the cached dependencies instead of downloading them every time. The cache is stored in `~/.npm` on Linux/macOS and is keyed based on the operating system, build type, and the hash of the `package-lock.json` file. This ensures that the cache is specific to the project and can be restored when needed.

see: https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#example-using-the-cache-action